Create MyAccountBtn component for user account actions
This commit introduces the 'MyAccountBtn' component, which provides a user account button with a dropdown menu for actions like accessing the account and logging out. The component uses user authentication data from the 'useAuthentication' hook and integrates with the 'Avatar' component for displaying the user's name and photo.

Key features of the 'MyAccountBtn' component:
- Displays user information with an avatar.
- Allows the user to open a dropdown menu with account-related actions.
- Supports the 'Logout' action, triggering the 'logoutUser' function from 'useAuthentication.'
- Enhances the user experience by providing easy access to account actions.

The 'MyAccountBtn' component contributes to the user interface and user account management in our application.